### 2020-02-16 #17 @rahulrajpl
[Re] Process Aware Stealthy Attack Detection on Industrial Control Systems.
**Original article:** Wissam Aoudi, Mikel Iturbe, and Magnus Almgren. 2018. Truth Will Out:
Departure-Based Process-Level Detection of Stealthy Attacks on Control
Systems. In 2018 ACM SIGSAC Conference on Computer and Communications
Security (CCS ’18), October 15–19, 2018, Toronto, ON, Canada. ACM, New
York, NY, USA, 15 pages. https://doi.org/10.1145/3243734.3243781

**Scientific domain:** Cybersecurity for Critical Infrastructure
