### 2020-02-18 #17 @rougier
Thank you for your submission. An editor will be soon assigned.

@ReScience/associate-editors Can someone edit this submission in Computer Science / Security (Python) ?
