### 2020-02-24 #17 @otizonaizit
Do we have anyone with well-founded knowledge in cybersecurity? It is out of my depth for sure…
