### 2020-02-24 #17 @otizonaizit
Also, I think I am not sure what is meant by

> **Suggested editor**: VSCode/ SublimeText

are the submitters aware of what is meant by _editor_ in a scientific paper submission?
