### 2020-02-26 #17 @rahulrajpl
Hi, all fellow ReScience members,

Let me submit that this paper does not require even base-level knowledge of cybersecurity. A good level of Linear Algebra and basic computer science knowledge is good enough to understand the paper and its reproduction. Original authors have applied the technique in the field of security of critical infrastructures.

Well, being an absolute newbie in technical writing, I am not sure about what is meant by 'Suggested editor'. I kindly request some 'veterans' to elaborate on it. If you mean the editor used to write the paper describing the reproduction, it is texStudio.

-regards, Rahul
