### 2020-03-09 #17 @AurelienNioche
I'm afraid it is very far from what I know, and I have no background
knowledge at all about cyber security...

On Mon, Mar 9, 2020 at 10:20 AM Nicolas P. Rougier <notifications@github.com>
wrote:

> @AurelienNioche <https://github.com/AurelienNioche> Can you review this
> submission ?
>
