### 2020-03-13 #17 @rougier
@ReScience/reviewers We need an extra reviewer for this replication in cyber-security (but no knowledge in cyber security needed, it's mostly about linea algebra and computer science).
See https://github.com/ReScience/submissions/issues/17
