### 2020-09-18 #17 @himito
Sorry @rougier and @eroesch, I was very busy and lost (then I forgot) due to the current situation and then holydays here in France.  Now I have the time to review the paper.
