### 2020-10-21 #17 @gabrielnetz
Sorry for the delay on pushing this, but after review everything seems to be working as intended, other then a small issue with a changed funcition on MatPlotLib that will become obsolete in a bit, so its probably worth revisiting. Other then that, everything seems good

[PyPASAD code report (1).pdf](https://github.com/ReScience/submissions/files/5415295/PyPASAD.code.report.1.pdf)
