### 2020-10-27 #17 @eroesch
@rahulrajpl Could you have a look at the minor comments that @gabrielnetz pointed out? I don't know if there is a recommended course of action, but basically we want to make sure your code is future proof.
