### 2021-05-29 #17 @eroesch
I PRed the updated metadata to the original repo. Haven’t heard back yet. Happy to try to generate the pdf myself. Should I?

Get Outlook for iOS<https://aka.ms/o0ukef>
