### 2021-05-30 #17 @eroesch
No, both @gabrielnetz and myself have been able to run the code. See @gabrielnetz's report in pdf earlier.
The only thing is that the repository does not contain the .tex, unfortunately. (@rahulrajpl)
