### 2021-05-31 #17 @himito
Dear all,

first, my sincere apologies for not having carried out the review earlier, hard times ...

I would also like to thank  @rahulrajpl for this submission. The code in the notebooks is clean and  self-documenting.

Concerning the document, my only comments are about the mix in the use of single and double quotations marks. Also, open quotations marks are not good, perhaps it's due to the use '' instead of `` in  LaTeX. Finally, the use of titlecase in all the titles of the steps would be good.

Concerning the code, I could reproduce the figures 😃. It is important to note that this is a **partial replication** as the document points out (two experiments of six were reproduced).

Concerning the git repository, a LICENSE file is missing, and also some description of each file/folder that it contains. For instance, how can we exploit the files inside the folder Exp. IV, Comparison with model based methods. Folder "1 - Scenario DA1" corresponds to which figure in the original document, etc. Finally, the README file does not present the instructions for running the experiments, which dependencies  do we need to install? which version of Python? etc.
