+++ Rescience Submission/article.py
@@ -1,0 +1,298 @@
+# ReScience yaml parser
+# Released under the BSD two-clauses licence
+
+import yaml
+
+class Contributor:
+    def __init__(self, role, name, orcid="", email="", affiliations=[]):
+        self.role = role
+        self.name = name
+        self.fullname = name
+        self.lastname = self.get_lastname(name)
+        self.abbrvname = self.get_abbrvname(name)
+        self.orcid = orcid
+        self.email = email
+        self.affiliations = affiliations
+
+    def get_abbrvname(self, name):
+        if not name: return ""
+        
+        if ',' in name:
+            lastname = name.split(",")[0]
+            firstnames = name.split(",")[1].strip().split(" ")
+        else:
+            lastname = name.split(" ")[-1]
+            firstnames = name.split(" ")[:-1]
+        abbrvname = ""
+        for firstname in firstnames:
+            if "-" in firstname:
+                for name in firstname.split("-"):
+                    abbrvname += name[0].strip().upper() + '.-'
+                abbrvname = abbrvname[:-1]
+            else:
+                abbrvname += firstname[0].strip().upper() + '.'
+        return abbrvname + " " + lastname
+
+
+    def get_lastname(self, name):
+        if not name: return ""
+        # Rougier, Nicolas P.
+        if ',' in name:
+            lastname = name.split(",")[0].strip()
+            firstname = name.split(",")[1].strip()
+        # Nicolas P. Rougier
+        else:
+            lastname = name.split(" ")[-1]
+            firstname = name.split(" ")[:-1]
+        return lastname
+    
+
+class Affiliation:
+    def __init__(self, code, name, address=""):
+        self.code = code
+        self.name = name
+        self.address = address
+
+class Repository:
+    def __init__(self, name, url, doi, swh=""):
+        self.name = name
+        self.url = url
+        self.doi = doi
+        self.swh = swh
+
+class Replication:
+    def __init__(self, cite, bib, url, doi):
+        self.cite = cite
+        self.bib = bib
+        self.url = url
+        self.doi = doi
+
+class Review:
+    def __init__(self, url, doi):
+        self.url = url
+        self.doi = doi
+
+class Date:
+    def __init__(self, date):
+        try:
+            import dateutil.parser
+
+            date = dateutil.parser.parse(date)
+            self.date = date
+            self.year = date.year
+            self.month = date.month
+            self.day = date.day
+            self.textual = self.date.strftime("%d %B %Y")
+        except:
+            import datetime
+            now = datetime.datetime.now()
+            self.date = now
+            self.year = now.year
+            self.month = now.month
+            self.day = now.day
+            self.textual = ""
+
+    def __str__(self):
+        return self.textual
+        #return self.date.strftime("%d %B %Y")
+
+    def __repr__(self):
+        return self.textual
+        # return self.date.strftime("%d %B %Y")
+        
+
+class Article:
+    def __init__(self, data):
+        self.title = ""
+        self.absract = ""
+        self.type = ""
+        self.domain = ""
+        self.language = ""
+        self.bibliography = ""
+        self.keywords = []
+        self.authors = []
+        self.editors = []
+        self.reviewers = []
+        self.affiliations = []
+        self.code = ""
+        self.data = ""
+        self.contact = ""
+
+        self.review = ""
+        self.replication = ""
+        
+        self.date_received = ""
+        self.date_accepted = ""
+        self.date_published = ""
+
+        self.journal_name = "" 
+        self.journal_issn = ""
+        self.journal_volume = ""
+        self.journal_issue = ""
+        self.article_number = ""
+        self.article_doi = ""
+        self.article_url = ""
+
+        self.parse(data)
+
[…]
