### 2021-06-28 #17 @eroesch
Hi @rahulrajpl, Thank you for this. Yes, I am happy to push this through on my own, if that helps. However, it looks like you only added the files of the template in your repository, *not* the files needed to recompile your article, such as the original figures, content.txt, bibliography.bib, etc. Any change you might have that at hand?
