### 2021-12-25 #17 @eroesch
Awesome. I’ll get in touch next week!
Could you push everything you have to your repository? Ideally I’d need to be able to recompile your article.tex from scratch.
