+++ Rescience Submission/template/bibliography.bib
@@ -1,0 +1,115 @@
+@inproceedings{Pasad:2018,
+author = {Aoudi, Wissam and Iturbe, Mikel and Almgren, Magnus},
+title = {Truth Will Out: Departure-Based Process-Level Detection of Stealthy Attacks on Control Systems},
+year = {2018},
+isbn = {9781450356930},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/3243734.3243781},
+doi = {10.1145/3243734.3243781},
+booktitle = {Proceedings of the 2018 ACM SIGSAC Conference on Computer and Communications Security},
+pages = {817–831},
+numpages = {15},
+keywords = {cyber-physical systems, departure detection, industrial control systems, partial isometry, intrusion detection, singular spectrum analysis, isometry trick, stealthy attacks},
+location = {Toronto, Canada},
+series = {CCS ’18}
+}
+  
+
+
+@article{Rougier:2018,
+  doi = {10.1038/d41586-018-04628-w},
+  year = {2018},
+  month = apr,
+  publisher = {Springer Nature},
+  volume = {556},
+  number = {7701},
+  pages = {309--309},
+  author = {Rougier, Nicolas P. and Hinsen, Konrad},
+  title = {Code reviewing puts extra demands on referees },
+  journal = {Nature},
+}
+
+@article{Science:2018,
+  doi = {10.1126/science.359.6377.725},
+  author = {Hutson, Matthew},
+  url = {http://science.sciencemag.org/content/359/6377/725},
+  year = {2018},
+  month = feb,
+  volume = {359},
+  number = {6377},
+  title = {Artificial intelligence faces a replication crisis},
+  journal = {Science},
+}
+
+@Misc{kaggle,
+  author = {Jordan D'Arcy},
+  title = {Introducing SSA for Time Series Decomposition},
+  url = {https://www.kaggle.com/jdarcy/introducing-ssa-for-time-series-decomposition},
+  year = {2018},
+  note = {Apache License, verison 2.0},
+  }
+
+@Misc{Roboto:2011,
+  author =       {Christian Robertson},
+  title =        {The Roboto family of fonts (Google)},
+  url =          {https://github.com/google/roboto},
+  year =         2011,
+  note =         {Apache License, verison 2.0},
+}
+
+@Misc{SourceSerifPro:2014,
+  author =       {Frank Grießhammer},
+  title =        {Source Serif Pro (Adobe Systems)},
+  url =          {https://github.com/adobe-fonts/source-serif-pro},
+  year =         2014,
+  note =         {SIL Open Font License, version 1.1},
+}
+
+@Misc{SourceCodePro:2012,
+  author =       {Paul D. Hunt},
+  title =        {Source Code Pro (Adobe Systems)},
+  url =          {https://github.com/adobe-fonts/source-code-pro},
+  year =         2012,
+  note =         {SIL Open Font License, version 1.1},
+}
+
+@article{Topalidou:2015,
+  author =       {Topalidou, Meropi and Rougier, Nicolas P.},
+  title =        {{[Re] Interaction between cognitive and motor cortico-basal
+                  ganglia loops during decision making: a computational study}},
+  journal =      {ReScience},
+  year =         2015,
+  volume =       1,
+  number =       1,
+  doi =          {10.5281/zenodo.27944},
+}
+
+@article{Rougier:2017,
+  doi =          {10.7717/peerj-cs.142},
+  author =       {Nicolas P. Rougier and Konrad Hinsen and Frédéric Alexandre
+                  and Thomas Arildsen and Lorena Barba and Fabien
+                  C. Y. Benureau and C. Titus Brown and Pierre de Buyl and Ozan
[…]
