+++ Rescience Submission/template/metadata.tex
@@ -1,0 +1,42 @@
+% DO NOT EDIT - automatically generated from metadata.yaml
+
+\def \codeURL{https://github.com/rahulrajpl/PyPASAD}
+\def \codeDOI{}
+\def \codeSWH{}
+\def \dataURL{}
+\def \dataDOI{}
+\def \editorNAME{}
+\def \editorORCID{}
+\def \reviewerINAME{}
+\def \reviewerIORCID{}
+\def \reviewerIINAME{}
+\def \reviewerIIORCID{}
+\def \dateRECEIVED{01 November 2018}
+\def \dateACCEPTED{}
+\def \datePUBLISHED{}
+\def \articleTITLE{[Re] Process Aware Stealthy Attack Detection in Industrial Control Systems}
+\def \articleTYPE{Reproduction}
+\def \articleDOMAIN{Cybersecurity of Critical Infrastructure}
+\def \articleBIBLIOGRAPHY{bibliography.bib}
+\def \articleYEAR{2020}
+\def \reviewURL{}
+\def \articleABSTRACT{Although the traditional IT security layer provided to Industrial Control Systems offers a great deal of protection, security managers often overlook the possibility of a potential stealthy attacks at the process level. In the paper, "Truth Will Out: Departure-Based Process-Level Detection of Stealthy Attacks on Control Systems" the authors bring out a novel approach to detecting stealthy attacks on the control system through a process-aware methodology which they call - PASAD. It enables early detection in the subtle variation of process behavior, thus averting strategic adversaries from maliciously manipulating the industrial process within the noise level. The original implementation for the PASAD algorithm was in Matlab. In this replication, I use Python, and popular visualization library, Matplotlib, to obtain the results claimed.}
+\def \replicationCITE{Wissam Aoudi, Mikel Iturbe, and Magnus Almgren. 2018. Truth Will Out: Departure-Based Process-Level Detection of Stealthy Attacks on Control Systems. In 2018 ACM SIGSAC Conference on Computer and Communications Security (CCS ’18), October 15–19, 2018, Toronto, ON, Canada. ACM, New York, NY, USA, 15 pages. https://doi.org/10.1145/3243734.3243781}
+\def \replicationBIB{}
+\def \replicationURL{}
+\def \replicationDOI{}
+\def \contactNAME{Rahul Raj}
+\def \contactEMAIL{rahulr@cse.iitk.ac.in}
+\def \articleKEYWORDS{python, matplotlib, time series analysis, Intrusion Detection, cybersecurity}
+\def \journalNAME{ReScience C}
+\def \journalVOLUME{4}
+\def \journalISSUE{1}
+\def \articleNUMBER{}
+\def \articleDOI{}
+\def \authorsFULL{Rahul Raj}
+\def \authorsABBRV{R. Raj}
+\def \authorsSHORT{Raj}
+\title{\articleTITLE}
+\date{}
+\author[1,\orcid{0000-0002-9049-1055}]{Rahul Raj}
+\affil[1]{Indian Institute of Technology, Kanpur, Uttar Pradesh (UP), India}
