### 2022-01-04 #17 @himito
Dear @rahulrajpl,

Thank you for updating the repository. However, many of my remarks regarding the  of the code repository have not been resolved. Could you take a look at them ? thanks

Best,
Jaime
