### 2022-01-17 #17 @eroesch
Thank @himito. To clarify, do I understand that the comments you refer to are (from your post 31/5/21):
- In the .tex manuscript, quote and double quotes are not used in a consistent way;
- The code repo should contain a README description of the structure of the notebooks;
- The code repo should contain a requirements.txt for libraries, incl. version, which should also be described in the README
- The code repo should contain a LICENSE specification;
- and some minor presentation tweaks in the notebooks (i.e. casing of titles)

There is nothing in the code proper, right? Your last comment is unclear, and I think both you and @gabrielnetz, and myself have managed to replicate the results presented.

As the editor, I am happy to make these changes (@rougier, @khinsen), and I'll also add a short description of the study itself in README, which I is missing, and that it is a full replication of only part of the original work. I am otherwise happy with the way the manuscript is presented, comparing the original matlab figures to the new python ones.
I timetabled it for this week, but I am having issues with the resci template itself and python code, which I'll have to troubleshoot first.
