### 2022-01-17 #17 @himito
Hi @eroesch,

Yep, the replication is OK.

My remarks are more focused on the code repository. For instance, have a good README, and a Licence of the code.
