### 2022-01-31 1844119
ENH: Clean up folders
