+++ Rescience Submission/_Trash/content.tex
@@ -1,6 +1,4 @@
 
-\subsection*{Introduction}
-
 Three years, we launched ReScience, a new scientific journal aimed at publishing the
 replication of existing computational research. Since ReScience published its first
 article\supercite{Topalidou:2015}, things have been
