+++ Rescience Submission/template/content.tex
@@ -1,172 +1,0 @@
-Although the traditional IT security layer provided to Industrial Control Systems offers a great deal of protection, security managers often overlook the possibility of a potential stealthy attacks at the process level. In the paper\supercite{Pasad:2018}, "Truth Will Out: Departure-Based Process-Level Detection of Stealthy Attacks on Control Systems" the authors bring out a novel approach to detecting stealthy attacks on the control system through a process-aware methodology which they call - PASAD. It enables early detection in the subtle variation of process behavior, thus averting strategic adversaries from maliciously manipulating the industrial process within the noise level. The original implementation for the PASAD algorithm was in Matlab. In this replication, I use Python, and popular visualization library, Matplotlib, to obtain the results claimed.
-
-\section*{Introduction}
-
-A typical control system utilizes sensors, actuators, and controllers to regulate some controlled process. Sensor devices measure some physical property and communicate the measurements to a controller (e.g., a PLC), which based on a control algorithm, correspondingly issues commands to actuators
-(e.g., control valves) that directly manipulate the physical process based on the received commands. A physical process may exhibit a noisy behavior by nature. A strategic adversary takes advantage of this fact and manipulates the process to stay within the noise level to remain undetected, thus leading to a potential degradation or suboptimal performance or, in the worst scenario, a complete failure of control system through cascading effect between various control loops. PASAD is capable of finding such attacks by utilizing a technique called 'singular spectrum analysis,' a model-free time-series analysis tool.
-
-\section*{Four Steps of PASAD}
-Consider a univariate real-valued time series of sensor measurements
-
-\begin{center}
-$\mathcal{T}=x_1, x_2,...,x_N, x_N+1$
-\end{center}
-
-\subsection*{Step 1: Embedding timeseries data into a trajectory matrix $X$}
-
-Let $L$ be an integer referred to as lag parameter, which is typically close to N/2. Initial $L$ points of $\mathcal{T}$ form the first column vector of Trajectory Matrix $X$. Then the lag vector moves to the next point, similar to the concept of a `moving window' until it covers all the points, forming corresponding column vectors for each. Final trajectory matrix of time series data is as follows:-
-
-$$
-X = \begin{bmatrix}
-	x_1 & x_2 & \dots  & x_K \\
-	x_2 & x_3 & \dots  & x_{K+1} \\
-	\vdots & \vdots & \ddots & \vdots \\
-	x_L & x_{L+1} & \dots  & x_N
-\end{bmatrix}
-$$
-
-\subsection*{Step 2: Singular Value Decomposition}
-
-
-To obtain the noise-reduced version of the above trajectory matrix, we perform singular value decomposition(SVD) on the matrix and obtain the L eigenvectors $u_1, u_2,...,u_L$ and their corresponding eigenvalues. From a scree plot, we determine the cumulative contribution of each eigenvalue to determine the top $r$ contributors to the time series. $r$ represents the statistical dimension, which is the number of degrees of freedom that account for the deterministic variability.
-
-\subsection*{Step 3: Creating a projection matrix}
-
-
[…]
