+++ Rescience Submission/article.tex
@@ -1,4 +1,7 @@
 \documentclass{rescience}
+\usepackage{graphicx}
+\usepackage{caption}
+\usepackage{subcaption}
 
 \begin{document}
 \input{header.tex}
