+++ Rescience Submission/metadata.yaml
@@ -1,133 +1,0 @@
-# To be filled by the author(s) at the time of submission
-# -------------------------------------------------------
-
-# Title of the article:
-#  - For a successful replication, it shoudl be prefixed with "[Re]"
-#  - For a failed replication, it should be prefixed with "[¬Re]"
-#  - For other article types, no instruction (but please, not too long)
-title: "[Re] Process Aware Stealthy Attack Detection in Industrial Control Systems"
-
-# List of authors with name, orcid number, email and affiliation
-# Affiliation "*" means contact author
-authors:
-  - name: Rahul Raj
-    orcid: 0000-0002-9049-1055
-    email: rahulr@cse.iitk.ac.in
-    affiliations: 1,*
-    
-  # - name: Charan
-    # orcid: 0000-0002-6972-589X
-    # email: Nicolas.Rougier@inria.fr
-    # affiliations:       # * is for contact author
-
-
-# List of affiliations with code (corresponding to author affiliations), name
-# and address. You can also use these affiliations to add text such as "Equal
-# contributions" as name (with no address).
-affiliations:
-  - code:    1
-    name:    Indian Institute of Technology, Kanpur
-    address: Uttar Pradesh (UP), India
-    
-  # - code:    2
-  #   name:    Synchrotron SOLEIL, Division Expériences
-  #   address: Gif sur Yvette, France
-    
-  # - code:    3
-  #   name:    INRIA Bordeaux Sud-Ouest
-  #   address: Bordeaux, France
-    
-  # - code:    4
-  #   name:    LaBRI, Université de Bordeaux, Institut Polytechnique de Bordeaux,
-  #            Centre National de la Recherche Scientifique, UMR 5800
-  #   address: Talence, France
-    
-  # - code:    5
-  #   name:    Institut des Maladies Neurodégénératives, Université  de Bordeaux,
-  #            Centre National de la Recherche Scientifique, UMR 5293
-  # #   address: Bordeaux, France
-
-
-# List of keywords (adding the programming language might be a good idea)
-keywords: python, matplotlib, time series analysis, Intrusion Detection, cybersecurity
-
-# Code URL and DOI/SWH (url is mandatory for replication, doi after acceptance)
-# You can get a DOI for your code from Zenodo, or an SWH identifier from
-# Software Heritage.
-#   see https://guides.github.com/activities/citable-code/
-code:
-  - url: https://github.com/rahulrajpl/PyPASAD
-  - doi: 
-  - swh: 
-
-# Data URL and DOI (optional if no data)
-data:
-  - url: 
-  - doi:
-
-# Information about the original article that has been replicated
-replication:
- - cite: "Wissam Aoudi, Mikel Iturbe, and Magnus Almgren. 2018. Truth Will Out:
-Departure-Based Process-Level Detection of Stealthy Attacks on Control
-Systems. In 2018 ACM SIGSAC Conference on Computer and Communications
-Security (CCS ’18), October 15–19, 2018, Toronto, ON, Canada. ACM, New
-York, NY, USA, 15 pages. https://doi.org/10.1145/3243734.3243781"# Full textual citation
- - bib:  # Bibtex key (if any) in your bibliography file
- - url:  # URL to the PDF, try to link to a non-paywall version
- - doi:  # Regular digital object identifier
-
-# Don't forget to surround abstract with double quotes
[…]
