### 2022-03-24 #17 @eroesch
Okeydok. After a bit of battle with the .tex and also the editing scripts. I am happy to finally close this submission. The article is published at https://zenodo.org/record/6375751
with citation:

Raj, R (2022).  [Re] Process Aware Stealthy Attack Detection in Industrial Control Systems. ReScience C. 7(1:4). doi: 10.5281/zenodo.6375751.

Well done, @rahulrajpl!
